[GEOT-4946] Fixes null geom when reading schema
FeatureJSON#readFeatureCollectionSchema does not support null geometry.
It now checks if the geometry is null.
Also removed check that forced primite types to be always the same when parsing entire JSON.

[GEOT-5026] Property named 'id' is recognized by FeatureJSON

FeatureJSON dropped property named 'id' when parsing GeoJSON

[GEOT-5045] FeatureJSON NPE getBounds fix

FeatureJSON#writeFeatureCollection, will fail when writing a feature collection that has a null geometry descriptor.
The exception is a NPE thrown at CollectionsDataStore#getBoundsInternal.
This has been reproduced in a test case by wrapping a feature collection in a feature source.